Move AES code inside sm_loader protection domain.
This was an implementation oversight. According to the Soteria design, AES code
should be included in the trusted SM_loader text section, such that its
untampered execution is (implicitely) attested through the
Sancus-module-specific and encryption key derivations.

The xtime function should also be included in the SM loader text section, but
now still runs in the unprotected domain to work around a compiler bug that
does not properly inline arithmetic multiplication code..